Retain :l, :xpos, etc. attributes on terminals in CGEL format (#38)
In edit mode, edits can be made in either the graphical tree or the CGEL bracketed text. When edits are made in the graphical tree the CGEL bracketing is regenerated. But the graphical tree does not encode attributes on nodes. This change caches all terminals from the CGEL bracketing, including any of their attributes. When graphical edits are made (no change to tokenization allowed), the updated tree is exported to CGEL format, then the cached terminals are swapped in so any attributes are retained, but keeping the category/function labels from the graphical tree. KNOWN ISSUE: attributes are retained only for terminal nodes; attributes on nonterminals may be lost.